Remove extras for backends, migrate to dependencies.yaml. (#10)
This PR changes to "bring your own backend" proposed by @vyasr by
removing the extras.

This PR also migrates all dependencies in pyproject.toml to be managed
by dependencies.yaml.

There is no longer anything specific to setuptools or scikit-build-core
in the repo. You could use any other backend and it would work just
fine. Also, extras don't work in a conda build environment. We prefer to
force the user to specify the backend and avoid hiding this as "magic."